requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai-whisper
transformers
torch
Pillow
